fix: autentication

+++ src/api/routes.py
@@ -13,8 +13,6 @@ from flask_bcrypt import Bcrypt
 from flask_mail import Mail, Message
 
 
-# TODO: Pasar token cada vez que un user o trainer quiera ver algo 
-
 api = Blueprint('api', __name__)
 CORS(api)  # Allow CORS requests to this API
 bcrypt = Bcrypt()
@@ -23,8 +21,19 @@ mail = Mail()
 
 @api.route('/users', methods=['GET'])
 @jwt_required()
-def handle_user_get():
-    pass
+def handle_users():
+    response_body = {}
+    current_user = get_jwt_identity()
+    if not current_user['role'] == 'administrators':
+        response_body['message'] = 'Not allowed!'
+        return response_body, 405
+    users = db.session.query(Users).all()
+    if not users:
+        response_body['message'] = 'No users currently registered'
+        return response_body,404
+    response_body['message'] = 'Users currently registered'
+    response_body['results'] = [single_user.serialize() for single_user in users]
+    return response_body, 200
 
 
 @api.route('/users', methods=['POST'])
@@ -61,8 +70,19 @@ def handle_signup_user():
 
 @api.route('/trainers', methods=['GET'])
 @jwt_required()
-def handle_trainers_get():
-    pass
+def handle_trainers():
+    response_body = {}
+    current_user = get_jwt_identity()
+    if not current_user['role'] == 'administrators':
+        response_body['message'] = 'Not allowed!'
+        return response_body, 405
+    trainers = db.session.query(Trainers).all()
+    if not trainers:
+        response_body['message'] = 'No trainers currently registered'
+        return response_body,404
+    response_body['message'] = 'Trainers currently registered'
+    response_body['results'] = [single_trainer.serialize() for single_trainer in trainers]
+    return response_body, 200
 
 
 @api.route('/trainers', methods=['POST'])
@@ -106,11 +126,23 @@ def handle_signup_trainer():
         return response_body, 200
 
 
-
 @api.route('/administrators', methods=['GET'])
 @jwt_required()
-def handle_admin_get():
-    pass
+def handle_admins():
+    response_body = {}
+    current_user = get_jwt_identity()
+    print(current_user['email'])
+    print(current_user['role'])
+    if not current_user['role'] == 'administrators':
+        response_body['message'] = 'Not allowed!'
+        return response_body, 405
+    admins = db.session.query(Administrators).all()
+    if not admins:
+        response_body['message'] = 'No administrators currently registered'
+        return response_body,404
+    response_body['message'] = 'Administrators currently registered'
+    response_body['results'] = [single_admin.serialize() for single_admin in admins]
+    return response_body, 200
 
 
 @api.route('/administrators', methods=['POST'])
@@ -118,6 +150,12 @@ def handle_admin_get():
 def handle_admin_signup():
     response_body = {}
     # TODO: Validar que en el token tenga un admin, si no es admin se retorna un 405
+    current_user = get_jwt_identity()
+    print(current_user['email'])
+    print(current_user['role'])
+    if not current_user['role'] == 'administrators':
+        response_body['message'] = 'Not allowed!'
+        return response_body, 405
     if request.method == 'POST':
         data = request.json
         if not data:
@@ -144,40 +182,47 @@ def handle_admin_signup():
         return response_body, 200
 
 
-# TODO: Hacer que el admin pueda crear specializaciones, aplicar token, para que las puedan ver entrenadores y admin
-@api.route('/specializations', methods=["GET", "POST"])
-def handle_specializations():
+@api.route('/specializations', methods=["POST"])
+@jwt_required()
+def add_specializations():
     response_body = {}
     specializations = db.session.query(Specializations).all()
-    if request.method == "GET":
-        if not specializations:
+    current_user = get_jwt_identity()
+    if not current_user['role'] == 'administrators':
+        response_body['message'] = 'Not allowed!'
+        return response_body, 405
+    data = request.json
+    if not data:
+        response_body["message"] = "No data provided"
+        return response_body, 400
+    if 'name' not in data:
+        response_body["message"] = "The 'name' field is required."
+        return response_body, 400
+    if any(specialization.name == data["name"].lower() for specialization in specializations):
+        response_body["message"] = "Specialization already exists"
+        return response_body, 400
+    new_specialization = Specializations(name=data["name"].lower(), 
+                                         description=data.get("description"), 
+                                         logo_url=data.get("logo_url"))
+    db.session.add(new_specialization)
+    db.session.commit()
+    response_body["message"] = "Specialization created"
+    response_body["specialization"] = new_specialization.serialize()
+    return response_body, 201
+    
+
+@api.route('/specializations', methods=['GET'])
+def get_specializations():
+    response_body = {}
+    specializations = db.session.query(Specializations).all()
+    if not specializations:
             response_body["message"] = "No specializations available"
             return response_body, 404
-        response_body["message"] = "Specializations available"
-        response_body["specializations"] = [specialization.serialize() for specialization in specializations]
-        return response_body, 200
-    if request.method == "POST":
-        data = request.json
-        if not data:
-            response_body["message"] = "No data provided"
-            return response_body, 400
-        if 'name' not in data:
-            response_body["message"] = "The 'name' field is required."
-            return response_body, 400
-        if any(specialization.name == data["name"].lower() for specialization in specializations):
-            response_body["message"] = "Specialization already exists"
-            return response_body, 400
-        new_specialization = Specializations(name=data["name"].lower(), 
-                                             description=data.get("description"), 
-                                             logo_url=data.get("logo_url"))
-        db.session.add(new_specialization)
-        db.session.commit()
-        response_body["message"] = "Specialization created"
-        response_body["specialization"] = new_specialization.serialize()
-        return response_body, 201
+    response_body["message"] = "Specializations available"
+    response_body["specializations"] = [specialization.serialize() for specialization in specializations]
+    return response_body, 200
 
 
-# Endpoints unico para los login
 @api.route('/login/<user_type>', methods=['POST'])
 def handle_login(user_type):
     response_body = {}
@@ -204,7 +249,8 @@ def handle_login(user_type):
             response_body['message'] = f'Wrong password for email {user.email}'
             return response_body, 401
         access_token = create_access_token(identity={"user": user.email,
-                                                     "role": user_type})
+                                                     "role": user_type,
+                                                     "id": user.id})
         response_body['message'] = 'Successfully logged in!'
         response_body['results'] = {"user": user.serialize(), 
                                     "role": user_type}
@@ -220,7 +266,8 @@ def handle_login(user_type):
             response_body['message'] = f'Wrong password for email {trainer.email}'
             return response_body, 401
         access_token = create_access_token(identity={"trainer": trainer.email,
-                                                     "role": user_type})
+                                                     "role": user_type,
+                                                     "id": trainer.id})
         response_body['message'] = 'Successfully logged in!'
         response_body['results'] = {"trainer": trainer.serialize(), 
                                     "role": user_type}
@@ -236,7 +283,8 @@ def handle_login(user_type):
             response_body['message'] = f'Wrong password for email {administrator.email}'
             return response_body, 401
         access_token = create_access_token(identity={"administrator": administrator.email,
-                                                     "role": user_type})
+                                                     "role": user_type,
+                                                     "id": administrator.id})
         response_body['message'] = 'Successfully logged in!'
         response_body['results'] = {"administrator": administrator.serialize(), 
                                     "role": user_type}
@@ -378,6 +426,10 @@ def handle_user_classes(id):
     user = Users.query.get(id)
     classes_user = UsersClasses.query.filter_by(user_id=id).all()
     classes_trainer = TrainersClasses.query.get(id)
+    current_user = get_jwt_identity()
+    if current_user['role'] not in ['administrators', 'users'] or current_user['id'] != user.id:
+        response_body["message"] = 'Not allowed!'
+        return response_body, 405
     if not user:
         response_body["message"] = "User not found"
         return response_body, 404
@@ -404,14 +456,12 @@ def handle_user_classes(id):
         if not trainer_class:
             response_body["message"] = "No Trainer class available"
             return response_body, 404
-        new_class = UsersClasses(
-            amount=data["amount"], 
-            stripe_status=data["stripe_status"], 
-            trainer_status="Pending", 
-            value=0,
-            user_id=id,
-            class_id=data["class_id"]
-        )
+        new_class = UsersClasses(amount=data["amount"], 
+                                 stripe_status= "Cart", 
+                                 trainer_status= "Pending", 
+                                 value=0,
+                                 user_id=id,
+                                 class_id=data["class_id"])
         db.session.add(new_class)
         db.session.commit()
         response_body["message"] = "Class added"
@@ -425,6 +475,7 @@ def handle_trainer_classes(id):
     response_body = {}
     trainer = Trainers.query.get(id)
     classes_trainer = TrainersClasses.query.filter_by(trainer_id=id).all()
+    current_user = get_jwt_identity()
     if not trainer:
         response_body["message"] = "Trainer not found"
         return response_body, 404
@@ -436,6 +487,9 @@ def handle_trainer_classes(id):
         response_body["Trainer classes"] = [class_trainer.serialize() for class_trainer in classes_trainer]
         return response_body, 200
     if request.method == "POST":
+        if (current_user['role'] == 'trainers' and current_user['id'] != trainer.id) or (current_user["role"] != "administrators"):
+            response_body["message"] = 'Not allowed!'
+            return response_body, 405
         data = request.json
         if not data:
             response_body["message"] = "No data provided"
@@ -444,16 +498,14 @@ def handle_trainer_classes(id):
         if existing_class:
             response_body["message"] = "Trainer class already exists for this datetime"
             return response_body, 400
-        new_trainer_class = TrainersClasses(
-            trainer_id=id, 
-            address=data["address"], 
-            capacity=data["capacity"], 
-            duration=data["duration"],
-            date=data["date"],
-            price=data["price"],
-            training_type=data["training_type"],
-            training_level=data["training_level"]
-        )
+        new_trainer_class = TrainersClasses(trainer_id=id, 
+                                            address=data["address"], 
+                                            capacity=data["capacity"], 
+                                            duration=data["duration"],
+                                            date=data["date"],
+                                            price=data["price"],
+                                            training_type=data["training_type"],
+                                            training_level=data["training_level"])
         db.session.add(new_trainer_class)
         db.session.commit()
         response_body["message"] = "New class create"
@@ -467,6 +519,7 @@ def handle_trainer_class(id, class_id):
     response_body = {}
     trainer = Trainers.query.get(id)
     trainer_class = TrainersClasses.query.filter(TrainersClasses.trainer_id == id, TrainersClasses.id == class_id).first()
+    current_user = get_jwt_identity()
     if not trainer: 
         response_body["message"] = "Trainer not found"
         return response_body, 404
@@ -477,35 +530,38 @@ def handle_trainer_class(id, class_id):
         response_body["message"] = "Trainer class"
         response_body["class"] = trainer_class.serialize()
         return response_body, 200
-    if request.method == "DELETE":
-        # No hace cancelar las classes si hay usuarios apuntados a ella
-        classes_user = UsersClasses.query.filter_by(class_id=class_id).all()
-        if classes_user:
-            response_body["message"] = "Unable to delete class, it has associated users"
-            return response_body, 400
-        db.session.delete(trainer_class)
-        db.session.commit()
-        response_body["message"] = "Class deleted successfully"
-        response_body["class"] = trainer_class.serialize()
-        return response_body, 200
-    if request.method == "PATCH":
-        data = request.json
-        if not data:
-            response_body["message"] = "No data provided for update"
-            return response_body, 400
-        if 'address' in data:
-            trainer_class.address = data["address"]
-        if 'duration' in data:
-            trainer_class.duration = data["duration"]
-        if 'date' in data:
-            trainer_class.date = data["date"]
-        if 'price' in data:
-            trainer_class.price = data["price"]
-        db.session.add(trainer_class)
-        db.session.commit()
-        response_body["message"] = "Class update"
-        response_body["result"] = trainer_class.serialize()
-        return response_body, 200
+    if (current_user['role'] == 'trainers' and current_user['id'] == trainer.id) or (current_user["role"] == "administrators"):
+        if request.method == "DELETE":
+            # No hace cancelar las classes si hay usuarios apuntados a ella
+            classes_user = UsersClasses.query.filter_by(class_id=class_id).all()
+            if classes_user:
+                response_body["message"] = "Unable to delete class, it has associated users"
+                return response_body, 400
+            db.session.delete(trainer_class)
+            db.session.commit()
+            response_body["message"] = "Class deleted successfully"
+            response_body["class"] = trainer_class.serialize()
+            return response_body, 200
+        if request.method == "PATCH":
+            data = request.json
+            if not data:
+                response_body["message"] = "No data provided for update"
+                return response_body, 400
+            if 'address' in data:
+                trainer_class.address = data["address"]
+            if 'duration' in data:
+                trainer_class.duration = data["duration"]
+            if 'date' in data:
+                trainer_class.date = data["date"]
+            if 'price' in data:
+                trainer_class.price = data["price"]
+            db.session.add(trainer_class)
+            db.session.commit()
+            response_body["message"] = "Class update"
+            response_body["result"] = trainer_class.serialize()
+            return response_body, 200
+    response_body["message"] = 'Not allowed!'
+    return response_body, 405
             
 
 @api.route('/users/<int:id>/classes/<int:class_id>', methods=["GET", "DELETE"])
@@ -518,6 +574,10 @@ def handle_user_class(id, class_id):
     user_class = UsersClasses.query.filter_by(user_id=id, class_id=class_id).first()
     # Buscar la clase del entrenador por ID de clase
     trainer_class = TrainersClasses.query.get(class_id)
+    current_user = get_jwt_identity()
+    if (current_user['role'] == 'users' and current_user['id'] != user.id) or (current_user["role"] != "administrators"):
+            response_body["message"] = 'Not allowed!'
+            return response_body, 405
     if not user:
         response_body["message"] = "User not found"
         return response_body, 404
@@ -539,20 +599,19 @@ def handle_user_class(id, class_id):
         return response_body, 200
 
 
-# New endpoint to GET and CREATE TrainersSpecializations
 @api.route('/trainers/<int:id>/specializations', methods=['GET','POST'])
 @jwt_required()
 def handle_trainers_specializations(id):
     response_body = {}
     trainer = db.session.query(Trainers).filter_by(id = id).first()
     if not trainer:
-        response_body['message'] = f'No trainer with trainer id {id} found!'
+        response_body['message'] = f'No trainer with trainer id {str(id)} found!'
         return response_body,404
     if request.method == 'GET':
         trainers_specializations = db.session.query(TrainersSpecializations).filter_by(trainer_id = id).all()
         if not trainers_specializations:
             response_body['message'] = 'No trainer specializations for trainer id ' + str(id)
-            return response_body,404
+            return response_body, 404
         response_body['message'] = 'Trainer specializations for trainer ' + str(id)
         response_body['results'] = [spec.serialize() for spec in trainers_specializations]
         return response_body,200
@@ -565,14 +624,13 @@ def handle_trainers_specializations(id):
         if not specialization:
             response_body['message'] = 'No specialization with id ' + data['specialization_id'] + ' !'
             return response_body, 404
-        new_trainer_specialization = TrainersSpecializations(
-            certification = data['certification'],
-            status = "Requested",
-            specialization_id = data['specialization_id'],  
-            trainer_id = id
-        )
+        new_trainer_specialization = TrainersSpecializations(certification = data['certification'],
+                                                             status = "Requested",
+                                                             specialization_id = data['specialization_id'],  
+                                                             trainer_id = id)
         db.session.add(new_trainer_specialization)
         db.session.commit()
         response_body['message'] = 'New specialization connected with trainer ' + str(id)
         response_body['results'] = new_trainer_specialization.serialize()
         return response_body,201
+
